backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic-settings>=2.1.0
redis>=5.0.1
python-multipart>=0.0.6
email-validator>=2.1.0
openai>=1.10.0
langchain>=0.1.0
qdrant-client>=1.7.0
websockets>=12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
stripe>=7.0.0
